Handle non-uniformly spaced cells in TsunamiScenario
The TsunamiScenario is now able to handle NetCDF input files with cells that are not uniformly spaced,
e.g. the distance between two neighboring cells does not have to be the same for all cells.
Additionally, both monotonically increasing and monotonically decreasing value orders for x and y dimensions
are now supported.